requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.3.0=mkl
absl-py=0.15.0=pyhd3eb1b0_0
astor=0.8.1=py36hecd8cb5_0
autopep8=1.6.0=pypi_0
blas=1.0=mkl
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.3.29=hecd8cb5_1
certifi=2021.5.30=py36hecd8cb5_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh4f3eec9_6
dlib=19.23.1=pypi_0
gast=0.2.2=py36_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.31.0=py36h7580e61_0
h5py=2.10.0=py36h0601b69_1
hdf5=1.10.6=hdbbcd12_0
importlib-metadata=4.8.1=py36hecd8cb5_0
imutils=0.5.4=pypi_0
intel-openmp=2022.0.0=hecd8cb5_3615
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=pypi_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libprotobuf=3.17.2=h2842e9f_1
markdown=3.3.4=py36hecd8cb5_0
matplotlib=3.3.4=pypi_0
mkl=2019.4=233
mkl-service=2.3.0=py36h9ed2024_0
mkl_fft=1.3.0=py36ha059aab_0
mkl_random=1.1.1=py36h959d312_0
ncurses=6.3=hca72f7f_2
numpy=1.19.2=py36h456fd55_0
numpy-base=1.19.2=py36hcfb5961_0
opencv-python=4.5.5.64=pypi_0
openssl=1.1.1n=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
pillow=8.4.0=pypi_0
pip=21.2.2=py36hecd8cb5_0
protobuf=3.17.2=py36h23ab428_0
pycodestyle=2.8.0=pypi_0
pyparsing=3.0.8=pypi_0
python=3.6.13=h88f2d9e_0
python-dateutil=2.8.2=pypi_0
readline=8.1.2=hca72f7f_1
scipy=1.5.2=py36h912ce22_0
setuptools=58.0.4=py36hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h707629a_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=mkl_py36ha38f243_0
tensorflow-base=2.0.0=mkl_py36h66b1bf0_0
tensorflow-estimator=2.0.0=pyh2649769_0
termcolor=1.1.0=py36hecd8cb5_1
tk=8.6.11=h7bc2e8c_0
toml=0.10.2=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.12.1=py36h1de35cc_1
xz=5.2.5=h1de35cc_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.12=h4dc903c_2
